Added new rightHandedRiemannSum function to utility.c
Simplified window_conv output to just include the size in KB and runtime in seconds

Added additional safety and debugging to waldpdf()

Finally removed the old threestage code

Converted twostage and threestage to use the new _ERROR_BOUND

Converted wald_bin to use the new rightHandedRiemannSum()

Fixed error in partitionLength calculation in all modes

Converted binned_conv and threestage_binconv to use the new rightHandedRiemannSum()

Removed lazy mode cruft from binned_conv

+++ IMT_FAST/imt_analysis.c
@@ -75,13 +75,11 @@ void IMT_analysis_April2017(const char *model, char * data_filename) {
 
 	printf("data[%d]={\n", data_size);
 	for (int i = 0; i < data_size; i++) {
-		printf("%f ", (float)data[i]);
+		printf("%.17e ", data[i]);
 		if (i % 8 == 7)
 			printf("\n");
 	}
 	printf("}\n");
-
-
 
 	configStruct config;
 
@@ -93,8 +91,6 @@ void IMT_analysis_April2017(const char *model, char * data_filename) {
     int onestagelagnomle = 0;
     int onestagefitnomle = 0;
     int threestagefitnomle = 0;
-   
-    
 
     const char *twostagefitnomle_str = "twostage";
     const char *onestagelagnomle_str = "onestagelag";
